Terminated a tough bug which resulted in zipping unwanted files (ZipmeTask)
It was caused by symlinks to directories. The symlink was detected as a file. It however contains directories which are going to be zipped. If such a directory contained an unwanted file (e.g. .DS_Store) it would not be checked and gets zipped. Fixed it with this commit (starting from line 161 of the new file). Also code style improvements.