link.php: Anticipate real folders to be replaced by symlinks
Until now when a real (not hard- or symlinked) folder existed at the location of a folder symlink the script would fail to delete it and, of course, create the symlink. Therefore it was impossible to replace a real folder with a symlink. With this commit a recursive deletion will take place in this case, allowing you to replace a real folder with a symlink, provided you have adequate rights to delete the real folder and its contents.

*Known issue*. Due to the way Windows processes some deletions you may have to run the link script _twice_ to get rid of real folders and replace them with symlinks.