Remove DocExport task
That was used in the Joomla! 1.5 days with DocImport 1.x. Back then we'd use DocExport to export the documentation as raw HTML pages inside a JPA archive at build time and DocImport to import the pages into com_content or K2 on the site. Since several years ago we import the XML file directly and transform it on the server, rendering this task obsolete. It's high time we did some spring cleaning!