Zipme: handle strange behavior of Phing filesets
Let's say you have a phing fileset with a basedir. That directory has two subdirectories and three files. You explicitly include two subdirectories and their contents and explicitly exclude the three files. What would you expect the fileset to include?

You'd reasonably expect it to include only the files in the two subdirectories you whitelisted.

Nope. It will also include the third subdirectory you neither included nor excluded. Because of that it will be listed as a (empty) folder item when you iterate the fileset. Even worse, it will be returned as a PhingFile object.

Guess what happens when you tell ZipArchive to addFile a directory? Nothing. Until you call close(). Then it barfs that the file is not a directory. AND SCREW YOU, DEVELOPER, WHO ARE TRYING TO DEBUG THIS because the actual error has occurred somewhere else. You will have no clue where.

Braindead PHP + braindead Phing. What a LOVELY combo...