common.xml: Use system xsltproc or PHP's xsltproc for documentation
When converting the DocBook XML documentation to PDF or ePub format you
can choose whether to use the xsltproc binary or the PHP's built-in
libxsl integration to do the conversion. The parameter to set is
`docs.php_xsltproc`. The default is 1, i.e. use PHP. This removes a hard
dependency to the libxsl-utils and replaces it with a PHP requirement.

`composer.json` has been updated with the PHP dependencies, including
PHP version and extensions, to prevent any mishaps ;)