common.xml: Build improvements
The new-release task now depends on the link and composer-install
tasks. This way, whenever you're building a new release, we make
sure that symlinks and Composer dependencies are set up properly.

Calling these tasks adds about 1 second to the build time but
prevents some of the most common errors when building a release
package. Therefore it's an acceptable trade-off.